Remove BOM and COGS generation examples
The functionality of these example scripts are redundant due to the
BOM and COGS custom actions. Every time we make a change to these, we
have to propagate that change to the actions repos as well. Further,
their presence here only increases the likelihood of confusion and
potential support burden.

See https://github.com/AllSpiceIO/py-allspice/issues/104