Eliminate N+1 API requests in issues
Fixes https://github.com/AllSpiceIO/py-allspice/issues/70.

This issue was caused by `Issue` fetching the repository every time it
is initialized via `parse_response`. This leads to an N+1 query,
which can be especially cumbersome when fetching all issues of a
repository with many issues. This commit fixes that by propagating the
repository from the caller of `Issue.parse_response`, and eager loading
it only in cases where one issue at a time is loaded.