Filter out blank components in list_components
For some formats we end up interpreting some non-component elements on
the board as components. These do not have any attributes and therefore
add a blank row to the bom or an empty object to list_components.

However, since it is possible that there is a user error that leads to
some components that shouldn't be blank end up being blank, a debug log
has been added to catch that case. The log respects the level of the
AllSpice client's logger, which is set to INFO by default.