Refactor BOM generation to centralize logic
As it turns out, System Capture BOM generation did not get the benefits
of the changes in #157 as the functions for filtering and sorting rows
were not called in the system capture BOM function. This was an entirely
avoidable mistake as the logic for BOM generation for all formats is
substantially the same. This commit moves the logic for BOM generation
into the `generate_bom` function and makes the format specific functions
aliases of the same. This should avoid similar mistakes in the future.