Add column configuration to skip in output
Adds an option, `skip_in_output`, to exclude a column from the output of
the BOM generation. This is useful when you want to group, filter or
sort by the value of the column but don't need it in the output. The
common use case for this is a property which indicates whether a
component should be excluded from the BOM.

Also changes an existing case to test this feature.