Create and use a log handler to actually set log level
Previously, the `log_level` argument to `AllSpice` didn't actually do
anything. Since the logger doesn't have a specified handler, it falls
back to the root logger which does not respect the result of
`setLogLevel`. This commit actually creates a handler for the allspice
client's logger, which enables `setLogLevel` to actually work.