Enforce max retries in list_components
Earlier, we would retry forever if the JSON wasn't generated. This
usually wasn't a problem because the JSON would eventually be generated
*or* lead to a server error, but we've seen cases where this hasn't been
true. Either way, we should've limited the number of retries,
`list_components` just didn't count the number of attempts. This commit
counts the attempts, stops at 5 retires, and tests this behavior.