Double default sleep in list_components
The default sleep time between retries is too short for orcad/system
capture files. This commit increases the sleep time to 1 second, giving
a total of roughly 4 seconds to time out. Ideally we'd make this a
configurable parameter, but that requires a lot more work and would also
involve generate-bom.