Add method to get and create Issue attachments
`Issue.assets` is often not set, like in the common case when
`Repository.get_issues` is used to fetch all issues. This meant that
getting the assets on the issue itself required re-loading it with
`Issue.request`, and there was no way to create an attachment on issue
issue. This commit fixes both problems.