Use a generic on `retry_generated`
`retry_generated.retry_not_yet_generated` can also work with
`get_generated_svg`, but the return type is specified as dict, which
leads to valid programs not type checking correctly. This commit changes
the return type to be generic based on the return type of the function
passed to `retry_not_yet_generated`.